refactor: rename AppVariants and move folder inside App
