+++ src/constants/routes.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import { AppCC, AppFC } from '@components/AppVariants';
+import { AppCC, AppFC } from '@components/App/AppRealization';
 
 const mainRoutes = [
   {
